Attach console handler to logger, fix bugs, move audio transcription to the speechxai_utils.py module
