+++ ferret/explainers/explanation_speech/equal_width/lime_equal_width_explainer.py
@@ -2,7 +2,6 @@ from typing import List
 from pydub import AudioSegment
 import numpy as np
 from ..lime_timeseries import LimeTimeSeriesExplainer
-from ..utils_removal import transcribe_audio
 from ..explanation_speech import ExplanationSpeech
 from ....speechxai_utils import pydub_to_np
