+++ pyproject.toml
@@ -48,7 +48,13 @@ pydub = { version = "0.25.1", optional = true }
 audiomentations = { version = "0.34.1", optional = true }
 audiostretchy = { version = "1.3.5", optional = true }
 pyroomacoustics = { version = "0.7.3", optional = true }
-whisperx = { version = "3.1.2", optional = true }
+
+# The version of WhisperX currently on PyPI has a problem with a dependency,
+# so the dependency needs to be installed from the GitHub repo, which in turns
+# prevents it from being used among the extras in pyproject.toml. Until a
+# working version of WhisperX is released, the users are required to install
+# it from the repo manually.
+# whisperx = { version = "3.1.2", optional = true }
 
 [tool.poetry.extras]
 speech = [
@@ -56,14 +62,14 @@ speech = [
     "audiomentations",
     "audiostretchy",
     "pyroomacoustics",
-    "whisperx"
+    # "whisperx"
 ]
 all = [
     "pydub",
     "audiomentations",
     "audiostretchy",
     "pyroomacoustics",
-    "whisperx"
+    # "whisperx"
 ]
